qcadoo/qcadoo-plugin/src/main/java/com/qcadoo/plugin/api/artifact/JarPluginArtifact.java
/**
 * ***************************************************************************
 * Copyright (c) 2010 Qcadoo Limited
 * Project: Qcadoo Framework
 * Version: 1.4
 *
 * This file is part of Qcadoo.
 *
 * Qcadoo is free software; you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation; either version 3 of the License,
 * or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty
 * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 * ***************************************************************************
 */
package com.qcadoo.plugin.api.artifact;

import java.io.File;
import java.io.FileInputStream;
import java.io.FileNotFoundException;
import java.io.InputStream;

/**
 * Plugin artifact build on JAR file.
 * 
 * @since 0.4.0
 */
public final class JarPluginArtifact implements PluginArtifact {

    private final File file;

    /**
     * Create artifact build on JAR file.
     * 
     * @param file
     *            JAR file
     * @throws IllegalStateException
     *             if file doesn't exists or cannot be read
     */
    public JarPluginArtifact(final File file) {
        if (!file.exists() || !file.canRead()) {
            throw new IllegalStateException("Cannot read file " + file.getAbsolutePath());
        }
        this.file = file;
    }

    @Override
    public String getName() {
        return file.getName();
    }

    @Override
    public InputStream getInputStream() {
        try {
            return new FileInputStream(file);
        } catch (FileNotFoundException e) {
            throw new IllegalStateException(e.getMessage(), e);
        }
    }

}
